requirements.txt
streamlit
openai
python-dotenv
chromadb
requests